invento-track-backend/requirements.txt
Flask-SQLAlchemy
Flask-Migrate
Flask
psycopg2